requirements.txt
streamlit
tensorflow
pandas
numpy
pillow
joblib
scikit-learn